코드 정리

+++ 2d_engine/SceneBegin.h
@@ -30,7 +30,7 @@ public:
 	virtual ~SceneBegin();
 
 	// IG2Scene
-	int		Type()						override { return EAPP_SCENE::EAPP_SCENE_BEGIN; }
+	int		Type()						override { return (int)EAPP_SCENE::EAPP_SCENE_BEGIN; }
 	int		Init(const std::any& ={})	override;
 	int		Destroy()					override;
 	int		Update(const std::any& t)	override;
