+++ 2d_engine/SceneEnd.h
@@ -37,7 +37,7 @@ public:
 	virtual ~SceneEnd();
 
 	// IG2Scene
-	int		Type()						override { return EAPP_SCENE::EAPP_SCENE_END; }
+	int		Type()						override { return (int)EAPP_SCENE::EAPP_SCENE_END; }
 	int		Init(const std::any& ={})	override;
 	int		Destroy()					override;
 	int		Update(const std::any& t)	override;
