+++ 2d_engine/SceneNone.h
@@ -32,7 +32,7 @@ public:
 	virtual ~SceneNone();
 
 	// IG2Scene
-	int		Type()						override { return EAPP_SCENE::EAPP_SCENE_NONE; }
+	int		Type()						override { return (int)EAPP_SCENE::EAPP_SCENE_NONE; }
 	int		Init(const std::any& ={})	override;
 	int		Destroy()					override;
 	int		Update(const std::any& t)	override;
