+++ 2d_engine/ScenePlay.h
@@ -38,7 +38,7 @@ public:
 	virtual ~ScenePlay();
 
 	// IG2Scene
-	int		Type()						override { return EAPP_SCENE::EAPP_SCENE_PLAY; }
+	int		Type()						override { return (int)EAPP_SCENE::EAPP_SCENE_PLAY; }
 	int		Init(const std::any& ={})	override;
 	int		Destroy()					override;
 	int		Update(const std::any& t)	override;
