+++ 2d_engine/testScene/SceneSpine.h
@@ -36,7 +36,7 @@ public:
 	virtual ~SceneSpine();
 
 	// IG2Scene
-	int		Type()						override { return EAPP_SCENE::EAPP_SCENE_SPINE; }
+	int		Type()						override { return (int)EAPP_SCENE::EAPP_SCENE_SPINE; }
 	int		Init(const std::any& ={})	override;
 	int		Destroy()					override;
 	int		Update(const std::any& t)	override;
